+++ src/polymer/polymer2-mixin-scanner.ts
@@ -61,8 +61,7 @@ class MixinVisitor implements Visitor {
       this._currentMixin = new ScannedPolymerElementMixin({
         name: namespacedName,
         sourceRange,
-        // TODO(justinfagnani): fix descriptions correctly in parseJsdoc
-        // description: docs.description,
+        description: parentJsDocs.description,
       });
       this._currentMixinNode = node;
       this._mixins.push(this._currentMixin);
@@ -82,8 +81,7 @@ class MixinVisitor implements Visitor {
       this._currentMixin = new ScannedPolymerElementMixin({
         name: namespacedName,
         sourceRange,
-        // TODO(justinfagnani): fix descriptions correctly in parseJsdoc
-        // description: docs.description,
+        description: nodeJsDocs.description,
       });
       this._currentMixinNode = node;
       this._mixins.push(this._currentMixin);
@@ -107,9 +105,8 @@ class MixinVisitor implements Visitor {
     const sourceRange = this._document.sourceRangeForNode(node);
     if (isMixin) {
       this._currentMixin = new ScannedPolymerElementMixin({
-          sourceRange,
-          // TODO(justinfagnani): fix descriptions correctly in parseJsdoc
-          // description: docs.description,
+        sourceRange,
+        description: docs.description,
       });
       this._currentMixinNode = node;
       this._mixins.push(this._currentMixin);
@@ -171,11 +168,8 @@ class MixinVisitor implements Visitor {
     if (mixin == null) {
       return;
     }
-    const comment = esutil.getAttachedComment(node) || '';
-    const docs = jsdoc.parseJsdoc(comment);
     const properties = getProperties(node, this._document);
 
-    mixin.description = docs.description ? docs.description.trim() : '';
     mixin.events = esutil.getEventComments(node);
     // mixin.sourceRange = this._document.sourceRangeForNode(node);
     if (properties) {