+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.0.0-alpha.28] - 2017-02-24
 * Support for `@memberof` jsdoc tag
+* Fix bad mixin descriptions
+
 
 ### Fixed
 * PackageUrlResolver encodes URIs returned from `resolve` method.